+++ src/AddWordForm.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { X } from 'lucide-react';
 
-const AddWordForm = React.memo(({ newWord, handleNewWordChange, handleExampleChange, addNewWord, setShowAddWordForm, emojiList }) => (
+const AddWordForm = React.memo(({ newWord, handleNewWordChange, handleExampleChange, addNewWord, setShowAddWordForm, emojiList, handleImageUpload }) => (
   <div className="fixed inset-0 bg-black/50 flex items-center justify-center z-50 p-4">
     <div className="bg-white rounded-2xl shadow-xl p-6 max-w-md w-full max-h-[90vh] overflow-y-auto">
       <div className="flex justify-between items-center mb-4">
