+++ src/data/initialWords.js
@@ -1,0 +1,89 @@
+export const initialWords = [
+  {
+    id: 1,
+    english: 'Adventure',
+    russian: 'Приключение',
+    category: 'Путешествия',
+    image: '🗺️',
+    examples: ['Life is an adventure', 'We went on an adventure'],
+    pronunciation: 'əd-ˈven-chər',
+    difficulty: 1,
+    nextReview: new Date().getTime(),
+    reviewCount: 0,
+    correctCount: 0,
+    starred: false,
+    status: 'new',
+    createdAt: new Date().getTime(),
+    lastReviewed: null
+  },
+  {
+    id: 2,
+    english: 'Mountain',
+    russian: 'Гора',
+    category: 'Природа',
+    image: '⛰️',
+    examples: ['The mountain is very high', 'We climbed the mountain'],
+    pronunciation: 'ˈmaʊn-tən',
+    difficulty: 1,
+    nextReview: new Date().getTime(),
+    reviewCount: 0,
+    correctCount: 0,
+    starred: false,
+    status: 'new',
+    createdAt: new Date().getTime(),
+    lastReviewed: null
+  },
+  {
+    id: 3,
+    english: 'Happiness',
+    russian: 'Счастье',
+    category: 'Эмоции',
+    image: '😊',
+    examples: ['Happiness is a choice', 'She found happiness in small things'],
+    pronunciation: 'ˈhæp-i-nəs',
+    difficulty: 2,
+    nextReview: new Date().getTime(),
+    reviewCount: 0,
+    correctCount: 0,
+    starred: false,
+    status: 'new',
+    createdAt: new Date().getTime(),
+    lastReviewed: null
+  },
+  {
+    id: 4,
+    english: 'Knowledge',
+    russian: 'Знание',
+    category: 'Образование',
+    image: '📚',
+    examples: ['Knowledge is power', 'He has vast knowledge'],
+    pronunciation: 'ˈnɒl-ɪdʒ',
+    difficulty: 2,
+    nextReview: new Date().getTime(),
+    reviewCount: 0,
+    correctCount: 0,
+    starred: false,
+    status: 'new',
+    createdAt: new Date().getTime(),
+    lastReviewed: null
+  },
+  {
+    id: 5,
+    english: 'Ocean',
+    russian: 'Океан',
+    category: 'Природа',
+    image: '🌊',
+    examples: ['The ocean is deep', 'We sailed across the ocean'],
+    pronunciation: 'ˈoʊ-ʃən',
+    difficulty: 1,
+    nextReview: new Date().getTime(),
+    reviewCount: 0,
+    correctCount: 0,
+    starred: false,
+    status: 'new',
+    createdAt: new Date().getTime(),
+    lastReviewed: null
+  }
+];
+
+export default initialWords;
