+++ src/data/shopItems.js
@@ -1,0 +1,68 @@
+export const shopItems = [
+  {
+    id: 'theme-dark',
+    name: 'Тёмная тема',
+    icon: '🌙',
+    price: 50,
+    type: 'theme',
+    description: 'Приятная для глаз тёмная тема'
+  },
+  {
+    id: 'theme-ocean',
+    name: 'Океанская тема',
+    icon: '🌊',
+    price: 75,
+    type: 'theme',
+    description: 'Освежающие океанские цвета'
+  },
+  {
+    id: 'theme-forest',
+    name: 'Лесная тема',
+    icon: '🌲',
+    price: 75,
+    type: 'theme',
+    description: 'Спокойные зелёные тона'
+  },
+  {
+    id: 'boost-xp',
+    name: 'XP Бустер',
+    icon: '⚡',
+    price: 100,
+    type: 'boost',
+    description: 'x2 XP на 24 часа'
+  },
+  {
+    id: 'boost-coins',
+    name: 'Монетный магнит',
+    icon: '🧲',
+    price: 150,
+    type: 'boost',
+    description: 'x2 монет на 24 часа'
+  },
+  {
+    id: 'hint-pack',
+    name: 'Пакет подсказок',
+    icon: '💡',
+    price: 30,
+    type: 'consumable',
+    description: '+5 подсказок для тренировок'
+  },
+  {
+    id: 'streak-freeze',
+    name: 'Заморозка серии',
+    icon: '🧊',
+    price: 200,
+    type: 'consumable',
+    description: 'Сохранить серию при пропуске дня'
+  },
+  {
+    id: 'unlock-category',
+    name: 'Новая категория',
+    icon: '📦',
+    price: 300,
+    type: 'unlock',
+    description: 'Разблокировать премиум категорию'
+  }
+];
+
+export default shopItems;
